requirements.txt
FastAPI
Flask
uvicorn
python-multipart
tensorflow
gdown
Pillow~=9.5.0
matplotlib
pandas
sqlalchemy
flask-restplus